remove the configure script
The configure scripts are GPL licensed. To be safe, we don't want to include software tools that use other licenses, that might conflict with our own. The cabal file now just adds the default CUDA installation paths for each platform. If that doesn't work, we still rely on the user to specify the appropriate paths with --extra-{include,lib}-dirs